Add files via upload
Procedural programming project done in Java meant to focus on user interaction and input validation/error handling. Calculates the amount of time it takes for a USF student to graduate (in years, months, and semesters), the credits needed to graduate, and the credits they have left before graduating. The program takes in the current year of the student, the number of credits they plan to take each specific semester, and the amount of credits already earned. The program also tells them if their current plan will allow them to graduate on time.